+++ packages/component/src/Flipper.js
@@ -10,18 +10,37 @@ const ROOT_CSS = css({
   touchAction: 'none'
 });
 
-export default ({ 'aria-label': ariaLabel, children, className, mode }) =>
+export default ({ 'aria-label': ariaLabel, children, className, keepFocus, mode }) => 
   <Context.Consumer>
-    { context =>
-      <button
-        aria-label={ ariaLabel || (mode === 'left' ? 'left' : 'right') }
-        className={ classNames(ROOT_CSS + '', className) }
-        onClick={ mode === 'left' ? context.scrollOneLeft : context.scrollOneRight }
-        type="button"
-      >
-        <div className="slider">
-          { children }
-        </div>
-      </button>
+    { ({ scrollOneLeft, scrollOneRight }) => {
+      let ref;
+
+      const click = () => {
+        (mode === 'left' ? scrollOneLeft : scrollOneRight)();
+        !keepFocus && ref.blur();
+      };
+
+      const keyDown = event => {
+        const { key } = event;
+        if (key === 'Enter' || key === ' ') {
+          event.preventDefault();
+          (mode === 'left' ? scrollOneLeft : scrollOneRight)();
+        }
+      };
+
+      return (
+        <button
+          aria-label={ ariaLabel || (mode === 'left' ? 'left' : 'right') }
+          className={ classNames(ROOT_CSS + '', className) }
+          onClick={click}
+          onKeyDown={keyDown}
+          type="button"
+          ref={element => ref = element}
+        >
+          <div className="slider">
+            { children }
+          </div>
+        </button>)
     }
+  }
   </Context.Consumer>