+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added `flipperKeepFocus` prop to `BasicFilm` and `keepFocus` prop to `Flipper` components, by [@tdurnford](https://github.com/tdurnford) in PR [#xx](https://github.com/spyip/react-film/pull/xx).
+
 ## [1.3.0] - 2019-10-24
 
 ### Added